+++ packages/beapi/src/events/events/index.ts
@@ -8,6 +8,29 @@ import { ItemUse } from './ItemUse.js'
 import { ItemUseOn } from './ItemUseOn.js'
 import { BlockPlaced } from './BlockPlaced.js'
 import { BlockDestroyed } from './BlockDestroyed.js'
+import { PlayerSwing } from './PlayerSwing.js'
+import { PlayerAttacked } from './PlayerAttacked.js'
+import { PlayerInViewVector } from './PlayerInViewVector.js'
+import { EntityAttacked } from './EntityAttacked.js'
+import { EntityInViewVector } from './EntityInViewVector.js'
+import { PlayerSleep } from './PlayerSleep.js'
+import { PlayerWake } from './PlayerWake.js'
+import { PlayerSneak } from './PlayerSneak.js'
+import { PlayerUnsneak } from './PlayerUnsneak.js'
+import { PlayerJump } from './PlayerJump.js'
+import { PlayerLanded } from './PlayerLanded.js'
+import { PlayerInWater } from './PlayerInWater.js'
+import { PlayerExitWater } from './PlayerExitWater.js'
+import { PlayerOnFire } from './PlayerOnFire.js'
+import { PlayerOffFire } from './PlayerOffFire.js'
+import { PlayerIsMoving } from './PlayerIsMoving.js'
+import { PlayerStoppedMoving } from './PlayerStoppedMoving.js'
+import { PlayerIsRiding } from './PlayerIsRiding.js'
+import { PlayerStoppedRiding } from './PlayerStoppedRiding.js'
+import { PlayerIsSprinting } from './PlayerIsSprinting.js'
+import { PlayerStoppedSprinting } from './PlayerStoppedSprinting.js'
+import { PlayerIsSwimming } from './PlayerIsSwimming.js'
+import { PlayerStoppedSwimming } from './PlayerStoppedSwimming.js'
 
 export const defaultEvents = [
   Tick,
@@ -20,4 +43,27 @@ export const defaultEvents = [
   ItemUseOn,
   BlockPlaced,
   BlockDestroyed,
+  PlayerSwing,
+  PlayerAttacked,
+  PlayerInViewVector,
+  EntityAttacked,
+  EntityInViewVector,
+  PlayerSleep,
+  PlayerWake,
+  PlayerSneak,
+  PlayerUnsneak,
+  PlayerJump,
+  PlayerLanded,
+  PlayerInWater,
+  PlayerExitWater,
+  PlayerOnFire,
+  PlayerOffFire,
+  PlayerIsMoving,
+  PlayerStoppedMoving,
+  PlayerIsRiding,
+  PlayerStoppedRiding,
+  PlayerIsSprinting,
+  PlayerStoppedSprinting,
+  PlayerIsSwimming,
+  PlayerStoppedSwimming,
 ]