+++ packages/beapi/src/player/Player.ts
@@ -9,11 +9,14 @@ import type { ExecuteCommandResponse, Health, Inventory, Location, Dimensions, G
 import { executeCommand } from '../command/executeCommand.js'
 import { players } from './PlayerManager.js'
 import { events } from '../events/EventManager.js'
+import type { Entity } from '..'
 
 export class Player {
   private readonly _name: string
   private _nameTag: string
   private _vanilla: MCPlayer
+  private _previousPlayerViewVector: Player | undefined
+  private _previousEntityViewVector: Entity | undefined
 
   public constructor(player: MCPlayer) {
     this._name = player.nameTag
@@ -174,5 +177,21 @@ export class Player {
       return 'nether'
     }
     return 'the end'
+  }
+
+  public setPreviousPlayerViewVector(player: Player | undefined): void {
+    this._previousPlayerViewVector = player
+  }
+
+  public getPreviousPlayerViewVector(): Player | undefined {
+    return this._previousPlayerViewVector
+  }
+
+  public setPreviousEntityViewVector(entity: Entity | undefined): void {
+    this._previousEntityViewVector = entity
+  }
+
+  public getPreviousEntityViewVector(): Entity | undefined {
+    return this._previousEntityViewVector
   }
 }