+++ packages/beapi/CHANGELOG.md
@@ -3,6 +3,47 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.3.0](https://github.com/MCBE-Utilities/BeAPI/compare/beapi-core@1.2.1...beapi-core@1.3.0) (2021-12-31)
+
+
+### Bug Fixes
+
+* **events:** entityattacked ([bd559f2](https://github.com/MCBE-Utilities/BeAPI/commit/bd559f23b72324dd5f10cfe175dd4352a4ae1025))
+* moved uuidv4 to utils ([39a2e50](https://github.com/MCBE-Utilities/BeAPI/commit/39a2e50b4b9c761e40cdd7a5811e57f321a1e6eb))
+* **player:** added getPreviousEntityViewVector ([84de6b8](https://github.com/MCBE-Utilities/BeAPI/commit/84de6b8f3bee84a5a76b9d3f73d92f2f07439213))
+* **types:** added PlayerSneak & PlayerUnsneak ([ebe5be6](https://github.com/MCBE-Utilities/BeAPI/commit/ebe5be6a168eee07bca7b50e3ac02da63d41a5f2))
+
+
+### Features
+
+* **events:** added EntityAttacked ([d533e62](https://github.com/MCBE-Utilities/BeAPI/commit/d533e624ee44bad02545a8cf6dd77fe2681d812b))
+* **events:** added EntityInViewVector ([561b1ee](https://github.com/MCBE-Utilities/BeAPI/commit/561b1ee80bca3b3d48b7d80788337600247ee33c))
+* **events:** added PlayerAttacked ([87edb52](https://github.com/MCBE-Utilities/BeAPI/commit/87edb528876b0c2c8058fc82cf611d896244f818))
+* **events:** added PlayerExitWater ([fa4c66a](https://github.com/MCBE-Utilities/BeAPI/commit/fa4c66a358b7f0d680ba3dbcc08adcc9236eed9c))
+* **events:** added PlayerInViewVector ([30b6589](https://github.com/MCBE-Utilities/BeAPI/commit/30b658979bfbfb34b632f033505231ee9eae212c))
+* **events:** added PlayerInWater ([fcc525b](https://github.com/MCBE-Utilities/BeAPI/commit/fcc525b60fbcd91b5d515ba7e14cc8b3657ad9ac))
+* **events:** added PlayerIsMoving ([d3d08ad](https://github.com/MCBE-Utilities/BeAPI/commit/d3d08ad080383602b3dda4309a525bfcbbf5067c))
+* **events:** added PlayerIsRiding ([4698cc9](https://github.com/MCBE-Utilities/BeAPI/commit/4698cc916a19ca9a2545d0ea3107db6c967829d9))
+* **events:** added PlayerIsSprinting ([1672a60](https://github.com/MCBE-Utilities/BeAPI/commit/1672a602a999f5954e2370197a63a9d97028f52f))
+* **events:** added PlayerIsSwimming ([d8e76f9](https://github.com/MCBE-Utilities/BeAPI/commit/d8e76f9e253897f7e3df277678c19a7fad4fd83a))
+* **events:** added PlayerJump ([67b5045](https://github.com/MCBE-Utilities/BeAPI/commit/67b5045c8cc803c57a69f853e905d54bbc146a6f))
+* **events:** added PlayerLanded ([6d90f19](https://github.com/MCBE-Utilities/BeAPI/commit/6d90f193d6a5e28db1ed39603be663e802904a57))
+* **events:** added PlayerOffFire ([fd4c609](https://github.com/MCBE-Utilities/BeAPI/commit/fd4c609281ca90d91dbeebc4fed174686d6318b7))
+* **events:** added PlayerOnFire ([ede0cb8](https://github.com/MCBE-Utilities/BeAPI/commit/ede0cb885186b24922221614ad5e13f4774c016f))
+* **events:** added PlayerSleep ([4ca1860](https://github.com/MCBE-Utilities/BeAPI/commit/4ca18602c0f9ed90572762afb849bdeaf1e1a6ff))
+* **events:** added PlayerSneak ([6d97f9a](https://github.com/MCBE-Utilities/BeAPI/commit/6d97f9a1738a965fc9f0d93e95b269d471bd7d5f))
+* **events:** added PlayerStoppedMoving ([9df00fc](https://github.com/MCBE-Utilities/BeAPI/commit/9df00fca438d8c161476fff576a2d21dc66cef0b))
+* **events:** added PlayerStoppedRiding ([a4bf238](https://github.com/MCBE-Utilities/BeAPI/commit/a4bf238267fbb7ef566e4e29d8e92363046e5262))
+* **events:** added PlayerStoppedSprinting ([9d3a01a](https://github.com/MCBE-Utilities/BeAPI/commit/9d3a01a8651b50fc543404c301ae68679d1d6a25))
+* **events:** added PlayerStoppedSwimming ([bb3e64b](https://github.com/MCBE-Utilities/BeAPI/commit/bb3e64badf474ec991a145d6539ed434064a2c38))
+* **events:** added PlayerSwing ([cab929a](https://github.com/MCBE-Utilities/BeAPI/commit/cab929a80770ba401e9fb920e6b5f9c74774e6c4))
+* **events:** added PlayerUnsneak ([73d43ae](https://github.com/MCBE-Utilities/BeAPI/commit/73d43ae2fd9d1debe2a305b42b861be517ea7fe5))
+* **events:** added PlayerWake ([bb07913](https://github.com/MCBE-Utilities/BeAPI/commit/bb07913b9285a380c115516a88a44ad67119cc30))
+
+
+
+
+
 ## [1.2.1](https://github.com/MCBE-Utilities/BeAPI/compare/beapi-core@1.2.0...beapi-core@1.2.1) (2021-12-30)